+++ src/dactyl_keyboard/param/tree/main.clj
@@ -97,8 +97,8 @@
      :validate [::valid/named-secondary-positions]}
     "A map where each item provides a name for a position in space. "
     "Such positions exist in relation to other named features of the keyboard "
-    "and can themselves be used as named features: Typically as supplementary "
-    "targets for `tweaks`, which are defined below.\n"
+    "and can themselves be used as anchors, most typically as supplementary "
+    "targets for `tweaks` (see below).\n"
     "\n"
     "An example:\n\n"
     "```secondaries:\n"
@@ -109,7 +109,8 @@
     "      segment: 2\n"
     "      offset: [1, 0, 0]\n"
     "    override [null, null, 2]\n"
-    "    translation: [0, 3, 0]\n```\n"
+    "    translation: [0, 3, 0]\n"
+    "    size: 4\n```\n"
     "\n"
     "This example gives the name `s0` to a point near some feature named "
     "`f0`, which must be defined elsewhere. All parameters in the `anchoring` "
@@ -121,9 +122,15 @@
     "of the secondary feature in the global vector space.\n"
     "\n"
     "In the example, `s0` is a position 1 mm to the local right of the "
-    "south-east corner of vertical segment 2 of `f0`, projected onto the "
+    "southeast corner of vertical segment 2 of `f0`, projected onto the "
     "global x-y plane at z = 2 (i.e. 2 mm above the floor), and then shifted "
-    "3 mm away from the user on that plane."]
+    "3 mm away from the user on that plane.\n"
+    "\n"
+    "The `size` parameter is relevant only as a fallback for `tweaks` that "
+    "target the position and do not specify a size in turn. If both are "
+    "omitted, tweaks fall back still further, to a tiny point.\n"
+    "\n"
+    "All of these properties are optional."]
    [:section [:main-body]
     "The main body of the keyboard is the main output of this application. "
     "It may be the only body. "